+++ lib/src/schema.dart
@@ -22,11 +22,27 @@ class Schema<T> {
   /// Validates and converts a list of JSON objects to a list of type T.
   ///
   /// Throws a [JsonTypeException] if the input is not a list.
-  List<T> many(dynamic json) {
-    if (json is! List) {
-      throw JsonTypeException(json, expected: "List<dynamic>", path: JsonPath.root());
+  List<T> list(dynamic json) {
+    if (json is! Iterable || json is String) {
+      throw JsonTypeException(json, expected: "Iterable (other than String)", path: JsonPath.root());
     }
     return [for (final (index, item) in json.indexed) validate(item, JsonPath.root()[index])];
+  }
+
+  /// Validates and converts a map of JSON objects to a map of type T.
+  ///
+  /// Throws a [JsonTypeException] if the input is not a map.
+  Map<String, T> map(dynamic json) {
+    if (json is! Map) {
+      throw JsonTypeException(json, expected: "Map", path: JsonPath.root());
+    }
+    late final Map<String, dynamic> jsonMap;
+    try {
+      jsonMap = json.cast<String, dynamic>();
+    } catch (e) {
+      throw JsonTypeException(jsonMap, expected: "Map<String, dynamic>", path: JsonPath.root());
+    }
+    return {for (final entry in jsonMap.entries) entry.key: validate(entry.value, JsonPath.root() / entry.key)};
   }
 
   /// Validates a JSON object against this schema and constructs an instance of T.
@@ -34,16 +50,21 @@ class Schema<T> {
   /// Throws [JsonTypeException] if the input is not a [Map<String, dynamic>].
   /// Throws [ArgumentErrorValidationException] if the constructor raises an ArgumentError.
   T validate(dynamic json, [JsonPath path = const JsonPath.root(), FieldInfo? field]) {
-    if (json is! Map<String, dynamic>) {
-      throw JsonTypeException(json, expected: "Map<String, dynamic>", path: path);
+    if (json is! Map) {
+      throw JsonTypeException(json, expected: "Map", path: path);
     }
-
-    final data = {for (final field in fields) field.name: field.value(json, path / field.name)};
+    late final Map<String, dynamic> jsonMap;
+    try {
+      jsonMap = json.cast<String, dynamic>();
+    } catch (e) {
+      throw JsonTypeException(jsonMap, expected: "Map<String, dynamic>", path: path);
+    }
+    final data = {for (final field in fields) field.name: field.value(jsonMap, path / field.name)};
 
     try {
       return constructor(data);
     } on ArgumentError catch (e) {
-      throw ArgumentErrorValidationException(json, e.message.toString(), field: field, path: path);
+      throw ArgumentErrorValidationException(jsonMap, e.message.toString(), field: field, path: path);
     }
   }
 }